Update SQL schema and function files

+++ add-comment-mentions-schema.sql
@@ -56,7 +56,7 @@ BEGIN
   SELECT
     ap.email,
     MAX(ap.name) as name, -- Take any non-null name
-    SUM(ap.count) as participation_count
+    SUM(ap.count)::BIGINT as participation_count
   FROM all_participants ap
   WHERE ap.email IS NOT NULL
     AND (
